// src/utils/models/diasafe.ts
export interface DiasafeModel{
date:string;
machine:string;

}
export interface MachineData {
    data: string;
    data1: string;
    data2: string;
    data3: string;
    data4: string;
    data5: string;
    data6: string;
    data7: string;
    data8: string;
    data9: string;
    data10: string;
    numeroMaquina: string;
    numeroMaquina1: string;
    numeroMaquina2: string;
    numeroMaquina3: string;
    numeroMaquina4: string;
    numeroMaquina5: string;
    numeroMaquina6: string;
    numeroMaquina7: string;
    numeroMaquina8: string;
    numeroMaquina9: string;
    numeroMaquina10: string;
  }
  